View構造分割でtype_body_length解消
